+++ src/parser/statements/udt_statement/udt_statement_parser.rs
@@ -6,7 +6,7 @@ pub enum UdtStatement {
     DropTypeStatement,
 }
 
-pub fn udt_statement(_lista: &mut [String], _index: i32) -> Result<Option<UdtStatement>, Error> {
+pub fn udt_statement(_lista: &mut [String], _index: usize) -> Result<Option<UdtStatement>, Error> {
     if let Some(_x) = create_type_statement(_lista, _index)? {
         return Ok(Some(UdtStatement::CreateTypeStatement));
     } else if let Some(_x) = alter_type_statement(_lista, _index)? {
@@ -19,21 +19,21 @@ pub fn udt_statement(_lista: &mut [String], _index: i32) -> Result<Option<UdtSta
 
 pub fn create_type_statement(
     _lista: &mut [String],
-    _index: i32,
+    _index: usize,
 ) -> Result<Option<UdtStatement>, Error> {
     Ok(None)
 }
 
 pub fn alter_type_statement(
     _lista: &mut [String],
-    _index: i32,
+    _index: usize,
 ) -> Result<Option<UdtStatement>, Error> {
     Ok(None)
 }
 
 pub fn drop_type_statement(
     _lista: &mut [String],
-    _index: i32,
+    _index: usize,
 ) -> Result<Option<UdtStatement>, Error> {
     Ok(None)
 }
